requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2

# Deep Learning & Embeddings
torch==2.0.1
sentence-transformers==2.2.2
transformers==4.33.2

# Recommendation Systems
implicit==0.6.1
surprise==1.1.3

# Data Warehouse
duckdb==0.9.1

# API Framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
schedule==1.2.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

